server/requirements.txt
absl-py==2.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
fonttools==4.55.0
fsspec==2024.10.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.68.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.2
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybboxes==0.1.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sahi==0.11.19
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.19.0
shapely==2.0.6
six==1.16.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.43
ultralyticsplus==0.0.28
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
